feat : `acceptance.step` 테스트 패키지 분리
